First pass at handling "between" operator for valueSource "field"
