+++ packages/react-querybuilder/src/query-builder.scss
@@ -80,3 +80,12 @@ $rqb-dnd-hover-border-bottom-width: 2px !default;
 [data-dnd='disabled'] .queryBuilder-dragHandle {
   display: none;
 }
+
+// Miscellaneous
+.rule {
+  .rule-value {
+    .rule-value-between-selector-1 {
+      margin-right: $rqb-spacing;
+    }
+  }
+}
