+++ packages/react-querybuilder/src/utils/formatQuery/defaultRuleProcessorJsonLogic.ts
@@ -1,6 +1,7 @@
 import type { JsonLogicVar } from 'json-logic-js';
+import { toArray } from '../../internal/toArray';
 import type { RQBJsonLogic, RuleType, ValueProcessorOptions } from '../../types/index.noReact';
-import { isValidValue, shouldRenderAsNumber, toArray } from './utils';
+import { isValidValue, shouldRenderAsNumber } from './utils';
 
 const convertOperator = (op: '<' | '<=' | '=' | '!=' | '>' | '>=') =>
   op
