+++ packages/react-querybuilder/src/utils/formatQuery/utils.ts
@@ -7,16 +7,6 @@ import type {
 
 export const numericRegex = /^\s*[+-]?(\d+|\d*\.\d+|\d+\.\d*)([Ee][+-]?\d+)?\s*$/;
 
-export const trimIfString = (val: any) => (typeof val === 'string' ? val.trim() : val);
-
-export const toArray = (v: any) =>
-  (Array.isArray(v)
-    ? v
-    : typeof v === 'string'
-    ? v.split(',').filter(s => !/^\s*$/.test(s))
-    : []
-  ).map(trimIfString);
-
 export const mapSQLOperator = (op: string) => {
   switch (op.toLowerCase()) {
     case 'null':
